requirements.txt
Django
djangorestframework
django-allauth
django-rest-auth
